fix changelog for a bug fix timeline
it looks like the bug fix for lookup-self http verb from POST to GET was done for 0.4.1 [GH-884]
my C#.NET client (Jan 20) was based off Vault's 0.4.1 and i was already using GET without any issues. 
it means this fix got in for 0.4.1 (the security build) and not 0.5.0.

again, please feel free to close the PR, if i am missing something.