Adding & to avoid errors.
I started a dev server. I stopped it and re-ran with &. Then I got errors and had to reinstall from scratch. This gave me some guidance so I thought I'd suggest that to prevent people from getting that error right off the bat by doing what I did in following the instructions verbatim.
If the output looks different, especially if the numbers are different
or the Vault is sealed, then restart the dev server and try again. The
only reason these would ever be different is if you're running a dev
server from going through this guide previously.