[server/listener_tcp] prevent binding 0.0.0.0 -> ::0
As an operator, it was surprising to see the
ipv6 bind address when configuring the ipv4 address.

This adds a special case for 0.0.0.0 to bind on tcp4
rather than the generic tcp. This more closely follows the
behavior seen in nginx and mysql.

The golang default of mapping 0.0.0.0 to ::0 makes sense as the
*default* case, we want to lean towards the *secure* case.

If an operating passes 0.0.0.0, we should honor it.
If they want dual stack, they can pass ::0.