ConfigureTLS() sets default HttpClient if nil
api.NewClient() creates a default HttpClient if one has not yet been
supplied, but ConfigureTLS() did not.

It seems like consistency here is useful: either client code need not
care whether the HttpClient was initialized (in this case, this is the
correct fix), or configuring the HttpClient is part of the contract for
initializing the client and api.NewClient() should also error if it is
not supplied.