Add acceptance test for chef-node backend
The test does require access to a Chef server.  A free hosted Chef
account works fine.  The test will create an environment, two roles, a
client and a node with two tags.  It checks that the node can be
authenticated and that the correct roles get assigned.  After it is
finished it removes the Chef objects that it created.