Add clients endpoint to bind policies to clients
Add a /clients endpoint that allows binding policy to specfic clients.
This is useful because tags, environments, and roles are all alterable
by a Chef client.  This means given a compromised client it could change
those attributes and gain access to secrets it was not intended to have.

The client is the only thing a compromised client can't change.  While
managing bindings to individual clients can be more work it is the most
secure and granular way to bind policy.