Have S3 group folders before sending response.
The ListObjectsV2 API allows you to specify a folder delimiter when
listing objects in a bucket. This makes S3 group the folders together
and only list the objects in the immediate prefix.

For prefixes with a large number of objects underneath, this can speed
up querying of intermediate paths significantly.