api/pki: there is no need for AllParallel method
It is sufficient to have only one List getter method which accepts the number of
goroutines as a parameter, because AllParallel(1) would be the same as All().
No need for channel, because the client can implement whatever it want after
getting the serialnumbers. All() is just a convenience method anyway.