api: Remove explicit http2 dependency
Commit 4156831 introduced an explicit dependency on `x/net/http2`, as it
there seemed to be an incompatibility between the builtin HTTP2 support
in the Go standard library and the external package.

This incompatibility is now seemingly fixed, and the explicit call to
`ConfigureTransport` is no longer required to enable HTTP2 upgrades in
the HTTP client.

By removing this explicit dependency on the `api` package, we greatly
simplify the vendoring story for other Go projects that depend on
`vault/api` to access Vault secrets. These projects will no longer need
to bring in all the `x/net/http2` package, which is a particularly heavy
dependency.