Fixed broken k8s TokenReview API link
Decided to point against 1.9 since it is most recent and likely to survive the longest.
The old link was redirecting to generic k8s 1.7 docs site with warning saying that k8s 1.7 docs are no longer being maintained. User could then click on "lastest version" link, but was not taken to newer version of the doc.

I don't think this API has changed since 1.7, so referring to 1.9 should be ok for customers on older versions of k8s. In any case, they can then figure out how to find older version of the API doc if they really want.