Add autocomplete for kv v2
This one was a bit tricky to implement because when I wrote the
autocompleter I made the assumption that Vault would never have such a
subcommand :).

The behavior is:

- `vault kv <cmd>` completes all mounts of _only_ type "kv" (v1 and v2)
- `vault <cmd>` completes all mounts _except_ type "kv" v2

This is technically a breaking change, but it doesn't make sense to
complete kv2 paths for generic Vault commands since they will fail.